Fixed the View classes according to the MVC structure.
MusicPlayerUI contains the main player of the app, it shows the play/pause, shuffle, repeat, next/previous, title, author and progress bar.

The Library tab was defined in LibraryView and the Playlist tab in PlaylistView. Both are called at the end of the MusicPlayerUI.

Also added img folder with some resources inside, later will be refractored to resources/img.